creating a nicer attribute API
