+++ sway-lsp/src/core/traverse_parse_tree.rs
@@ -4,9 +4,7 @@ use crate::{
     utils::token::{desugared_op, to_ident_key, type_info_to_symbol_kind},
 };
 use sway_core::{
-    constants::{
-        DESTRUCTURE_PREFIX, MATCH_RETURN_VAR_NAME_PREFIX, TUPLE_NAME_PREFIX,
-    },
+    constants::{DESTRUCTURE_PREFIX, MATCH_RETURN_VAR_NAME_PREFIX, TUPLE_NAME_PREFIX},
     parse_tree::{Literal, MethodName},
     type_system::{TypeArgument, TypeParameter},
     AbiCastExpression, ArrayIndexExpression, AstNode, AstNodeContent, CodeBlock, Declaration,
