+++ sway-lsp/src/server.rs
@@ -81,7 +81,6 @@ fn capabilities() -> ServerCapabilities {
         }),
         // document_formatting_provider: Some(OneOf::Left(true)),
         definition_provider: Some(OneOf::Left(true)),
-        hover_provider: Some(HoverProviderCapability::Simple(true)),
         ..ServerCapabilities::default()
     }
 }
