+++ sway-lsp/src/utils/attributes.rs
@@ -3,19 +3,13 @@ use crate::{
     core::token::{AstToken, TokenMap},
     utils::token::to_ident_key,
 };
-use std::collections::HashMap;
 use sway_ast::AttributeDecl;
 use sway_core::{
+    attributes_to_map,
     constants::{DOC_ATTRIBUTE_NAME, STORAGE_PURITY_ATTRIBUTE_NAME},
-    Declaration,
+    Attribute, Declaration,
 };
 use sway_types::Ident;
-
-#[derive(Clone, Debug)]
-pub struct Attribute {
-    pub name: Ident,
-    pub args: Vec<Ident>,
-}
 
 pub(crate) fn attribute_decls(decl_ident: &Ident, tokens: &TokenMap) -> Option<Vec<AttributeDecl>> {
     if let Some(item) = tokens.get(&to_ident_key(decl_ident)) {
@@ -50,29 +44,29 @@ pub(crate) fn storage_attributes(decl_ident: &Ident, tokens: &TokenMap) -> Optio
     })
 }
 
-pub(crate) fn attributes_to_map(attribute_list: &[AttributeDecl]) -> HashMap<String, Vec<Attribute>> {
-    let mut attrs_map: HashMap<String, Vec<Attribute>> = HashMap::new();
-    for attr_decl in attribute_list {
-        let attr = attr_decl.attribute.get();
-        let args = attr
-            .args
-            .as_ref()
-            .map(|parens| parens.get().into_iter().cloned().collect())
-            .unwrap_or_else(Vec::new);
+// pub(crate) fn attributes_to_map(attribute_list: &[AttributeDecl]) -> HashMap<String, Vec<Attribute>> {
+//     let mut attrs_map: HashMap<String, Vec<Attribute>> = HashMap::new();
+//     for attr_decl in attribute_list {
+//         let attr = attr_decl.attribute.get();
+//         let args = attr
+//             .args
+//             .as_ref()
+//             .map(|parens| parens.get().into_iter().cloned().collect())
+//             .unwrap_or_else(Vec::new);
 
-        let attribute = Attribute {
-            name: attr.name.clone(),
-            args,
-        };
-        let name = attr.name.as_str();
-        match attrs_map.get_mut(name) {
-            Some(old_args) => {
-                old_args.push(attribute);
-            }
-            None => {
-                attrs_map.insert(name.to_string(), vec![attribute]);
-            }
-        }
-    }
-    attrs_map
-}
+//         let attribute = Attribute {
+//             name: attr.name.clone(),
+//             args,
+//         };
+//         let name = attr.name.as_str();
+//         match attrs_map.get_mut(name) {
+//             Some(old_args) => {
+//                 old_args.push(attribute);
+//             }
+//             None => {
+//                 attrs_map.insert(name.to_string(), vec![attribute]);
+//             }
+//         }
+//     }
+//     attrs_map
+// }
